[#65]: Fix onload handler script
