+++ src/pshry.com/assets/js/main.js
@@ -1,19 +1,14 @@
+import { fireHandlers } from './onload'
+
 /**
- * Fire stylesheet onload handlers if the CSP has blocked inline scripts.
+ * Initialize main script.
  *
  * @since unreleased
  */
-function fireStylesheetOnloadHandlers () {
-	const links = document.querySelectorAll(
-		'[media="print"][onload="this.media=\'all\'"]'
-	)
-	if (!links) return
-
-	for (let i = 0; i < links.length; i++) {
-		const link = links[i]
-		link.setAttribute('media', 'all')
-		link.removeAttribute('onload')
-	}
+function init () {
+	// Fire onload handlers.
+	fireHandlers()
 }
 
-fireStylesheetOnloadHandlers()
+// Initialize main script.
+init()
